Abstract a mq_agent module
code fix after rebase
